Revert "Clean ups and refactoring pre-work 2"
This reverts commit b28fc68b71d2c86f123dfbe330ea9c9e34c9e492.
