+++ tests/test_analysis_utils.py
@@ -4,16 +4,6 @@ from kbmod.analysis_utils import *
 from kbmod.search import *
 
 class test_analysis_utils(unittest.TestCase):
-
-    def make_trajectory_(self, x0, y0, xv, yv, lh):
-        t = trajectory()
-        t.x = x0
-        t.y = y0
-        t.x_v = xv
-        t.y_v = yv
-        t.lh = lh
-        return t
-
     def setUp(self):
         # The configuration parameters.
         self.default_mask_bits_dict = {
@@ -411,37 +401,6 @@ class test_analysis_utils(unittest.TestCase):
         self.assertIsNotNone(res["stamps"])
         self.assertIsNotNone(res["final_results"])
 
-    def test_clustering(self):
-        cluster_params = {}
-        cluster_params["x_size"] = self.dim_x
-        cluster_params["y_size"] = self.dim_y
-        cluster_params["vel_lims"] = [self.min_vel, self.max_vel]
-        cluster_params["ang_lims"] = [self.min_angle, self.max_angle]
-        cluster_params["mjd"] = np.array(self.stack.get_times())
-
-        trjs = [self.make_trajectory_(10, 11, 1, 2, 100.0),
-                self.make_trajectory_(10, 11, 10, 20, 100.0),
-                self.make_trajectory_(40, 5, -1, 2, 100.0),
-                self.make_trajectory_(5, 0, 1, 2, 100.0),
-                self.make_trajectory_(5, 1, 1, 2, 100.0),
-               ]
-
-        # Try clustering with positions, velocities, and angles.
-        self.config["cluster_type"] = "all"
-        self.config["eps"] = 0.1
-        kb_post_process = PostProcess(self.config)
-        keep = kb_post_process.gen_results_dict()
-        keep["results"] = trjs
-        results_dict = kb_post_process.apply_clustering(keep, cluster_params)
-        self.assertEqual(len(results_dict["results"]), 4)
- 
-        # Try clustering with only positions.
-        self.config["cluster_type"] = "position"
-        kb_post_process = PostProcess(self.config)
-        keep = kb_post_process.gen_results_dict()
-        keep["results"] = trjs
-        results_dict = kb_post_process.apply_clustering(keep, cluster_params)
-        self.assertEqual(len(results_dict["results"]), 3)
 
 if __name__ == "__main__":
     unittest.main()
